triangleWithBarycentric for model

+++ model.cpp
@@ -73,7 +73,7 @@ void Model::load_texture(std::string filename, const std::string suffix, TGAImag
 
 vec3 Model::normal(const vec2 &uvf) const {
     TGAColor c = normalmap.get(uvf[0]*normalmap.width(), uvf[1]*normalmap.height());
-    return vec3{(double)c[2],(double)c[1],(double)c[0]}*2./255. - vec3{1,1,1};
+    return vec3{(float)c[2],(float)c[1],(float)c[0]}*2./255. - vec3{1.,1.,1.};
 }
 
 vec2 Model::uv(const int iface, const int nthvert) const {
